Node_JS_tutorial/old_qlts_project/OLD/daidv_requirement.txt
aioredis==1.3.1
#asgi-redis==1.4.3
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.3.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.5.7
bcrypt==4.0.1
cached-property==1.5.2
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.1
constantly==15.1.0
country-list==1.0.0
cryptography==3.4.7
cssselect2==0.4.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.12
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.10
django-admin-search==0.3.7
django-base64field==1.0
django-braces==1.14.0
django-ckeditor==6.2.0
django-filter==21.1
django-import-export==2.5.0
django-ipware==4.0.2
django-js-asset==1.2.2
django-model-utils==4.1.1
django-netfields==1.3.0
django-oauth-toolkit==1.5.0
django-oauth2-provider==0.2.6.1
django-postgres-extra==2.0.6
django-redis==5.2.0
django-rest-framework-social-oauth2==1.1.0
django-user-agents==0.4.0
et-xmlfile==1.1.0
filetype==1.0.7
gevent==22.10.1
greenlet==1.1.3.post0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.12.0
incremental==21.3.0
Jinja2==3.0.3
jwcrypto==0.9.1
lxml==4.6.4
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.4
msgpack-python==0.5.6
multidict==5.1.0
netaddr==0.8.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
Pillow==8.3.1
psutil==5.9.2
psycopg2-binary==2.8
py-avataaars==1.1.2
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycountry==22.3.5
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.4
python-docx==0.8.11
python-magic==0.4.24
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.4.0
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
shortuuid==1.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.1
toml==0.10.2
Twisted==22.8.0
txaio==21.2.1
typing-extensions==3.10.0.2
ua-parser==0.16.1
Unidecode==1.3.2
Unipath==1.1
urllib3==1.26.6
user-agents==2.2.0
webencodings==0.5.1
websocket==0.2.1
websockets==10.3
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0

PyJWT==2.6.0
djangorestframework==3.12.4
#djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
python-dotenv==0.21.0
django-multiselectfield==0.1.12